Formatted for NeadWerx.

+++ Arduino/src/circnoise_pal_1.h
@@ -11,7 +11,7 @@
 // moves a noise up and down while slowly shifting to the side
 void circnoise_pal_1() {
 
-  uint8_t scale = 1000;                               // the "zoom factor" for the noise
+  int scale = 1000;                               // the "zoom factor" for the noise
 
   for (uint16_t i = 0; i < NUM_LEDS_PER_STRIP; i++) {
