+++ Arduino/src/circnoise_pal_3.h
@@ -11,7 +11,7 @@
 // no x/y shifting but scrolling along z
 void circnoise_pal_3() {
 
-  uint8_t scale = 1000;                               // the "zoom factor" for the noise
+  int scale = 1000;                               // the "zoom factor" for the noise
 
   for (uint16_t i = 0; i < NUM_LEDS_PER_STRIP; i++) {
